adding properties for order refunds
