+++ tap_shopify/streams/base.py
@@ -256,123 +256,6 @@ class Stream():
         for obj in self.get_objects():
             yield obj.to_dict()
 
-    def get_children_by_graph_ql(self, child, child_parameters):
-        LOGGER.info("Getting data with GraphQL")
-
-        updated_at_min = self.get_bookmark()
-
-        stop_time = singer.utils.now().replace(microsecond=0)
-        date_window_size = float(Context.config.get("date_window_size", DATE_WINDOW_SIZE))
-        results_per_page = Context.get_results_per_page(RESULTS_PER_PAGE)
-
-        # Page through till the end of the resultset
-        while updated_at_min < stop_time:
-            after = None
-            updated_at_max = updated_at_min + datetime.timedelta(days=date_window_size)
-
-            if updated_at_max > stop_time:
-                updated_at_max = stop_time
-            singer.log_info("getting from %s - %s", updated_at_min,
-                            updated_at_max)
-            while True:
-                query = self.get_graph_query(updated_at_min,
-                                             updated_at_max,
-                                             150,
-                                             child,
-                                             child_parameters,
-                                             results_per_page,
-                                             after=after)
-                with metrics.http_request_timer(self.name):
-                    data = self.excute_graph_ql(query)
-
-                data = data[self.name]
-                page_info = data['pageInfo']
-                edges = data["edges"]
-                for edge in edges:
-                    after = edge["cursor"]
-                    node = edge["node"]
-                    yield node
-                if not page_info["hasNextPage"]:
-                    Context.state.get('bookmarks', {}).get(self.name, {}).pop('since_id', None)
-                    self.update_bookmark(utils.strftime(updated_at_max + datetime.timedelta(seconds=1)))
-                    break
-
-            updated_at_min = updated_at_max + datetime.timedelta(seconds=1)
-
-    @shopify_error_handling
-    def excute_graph_ql(self, query):
-        try:
-            # the execute function sometimes prints and this causes errors for the target, so I block printing for it
-            with HiddenPrints():
-                response = json.loads(shopify.GraphQL().execute(query))
-        except Exception:
-            raise GraphQLGeneralError("Execution failed", code=500)
-
-        if 'data' in response:
-            return response['data']
-        else:
-            if "errors" in response:
-                errors = response["errors"]
-                singer.log_info(errors)
-                if errors[0]["extensions"]["code"] == "THROTTLED":
-                    raise GraphQLThrottledError("THROTTLED", code=429)
-
-            raise GraphQLGeneralError("Failed", code=500)
-
-    def get_graph_query(self, created_at_min, created_at_max, limit, child, child_parameters, child_limit=100,
-                        after=None):
-        query = """{
-                      orders(first:%i %s ,query:"created_at:>'%s' AND created_at:<'%s'") {
-                        pageInfo { # Returns details about the current page of results
-                          hasNextPage # Whether there are more results after this page
-                          hasPreviousPage # Whether there are more results before this page
-                        }
-                        edges{
-                          cursor
-                          node{
-                            id,
-                            %s(first:%i){
-                              %s
-                            }
-                            createdAt
-                          }
-                        }
-                      }
-                }"""
-        after_str = ''
-        if after:
-            after_str = ',after:"%s"' % after
-        query = query % (limit, after_str, created_at_min, created_at_max, child, child_limit, child_parameters)
-        return query
-
-    def get_table_schema(self):
-        streams = Context.catalog["streams"]
-        schema = None
-        for stream in streams:
-            if stream["tap_stream_id"] == self.name:
-                schema = stream["schema"]
-                break
-
-        return schema
-
-    def get_graph_ql_prop(self, schema):
-        properties = schema["properties"]
-        ql_fields = []
-        for prop in properties:
-            if "generated" in properties[prop]["type"]:
-                continue
-            if 'object' in properties[prop]['type']:
-                if properties[prop]["properties"]:
-                    ql_field = "%s{%s}" % (prop, self.get_graph_ql_prop(properties[prop]))
-                    ql_fields.append(ql_field)
-            elif 'array' in properties[prop]['type']:
-                if properties[prop]["items"]["properties"]:
-                    ql_field = "%s{%s}" % (prop, self.get_graph_ql_prop(properties[prop]["items"]))
-                    ql_fields.append(ql_field)
-            else:
-                ql_fields.append(prop)
-        return ','.join(ql_fields)
-
     @shopify_error_handling
     def get_next_page(self, resource_find):
         return resource_find.next_page()
