+++ tap_shopify/streams/order_refunds.py
@@ -3,6 +3,7 @@ import shopify
 from tap_shopify.context import Context
 
 from tap_shopify.streams.graph_ql_stream import GraphQlChildStream
+import json
 
 
 class OrderRefunds(GraphQlChildStream):
@@ -11,6 +12,13 @@ class OrderRefunds(GraphQlChildStream):
     parent_key_access = "refunds"
     parent_name = "orders"
     parent_id_ql_prefix = 'gid://shopify/Order/'
+    child_per_page = 5
+    parent_per_page = 15
+    need_edges_cols = ["refundLineItems"]
+
+    def transform_obj(self, obj):
+        obj["refundLineItems"] = [trans_obj["node"] for trans_obj in obj["refundLineItems"]["edges"]]
+        return obj
 
 
 Context.stream_objects['order_refunds'] = OrderRefunds
